add trigger limit
git-svn-id: https://svn.ego-gw.it/svn/advsw/Omicron/trunk@59564 47e613b6-93f3-4244-b775-e06b1756610f

+++ src/Omap.cc
@@ -110,7 +110,7 @@ void Omap::SetBins(const double aQ, const double aFrequencyMin, const double aFr
     tiletag[f] = new bool [Nt];
     for(int t=0; t<Nt; t++) tiletag[f][t]=true;
 
-    Ntiles+=Nt;
+    Ntiles+=(long int)Nt;
   }
 
   return;
